created seeders, fix/update migrations

+++ migrations/20200512012036-create-produto.js
@@ -6,7 +6,7 @@ module.exports = {
         allowNull: false,
         autoIncrement: true,
         primaryKey: true,
-        type: Sequelize.INTEGER,
+        type: Sequelize.INTEGER(1),
       },
       produto: {
         type: Sequelize.STRING(45),
